LekzyTTP-SMS-GITHUB-COMPLETE-FINAL/requirements.txt
bcrypt==4.0.1
flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-telegram-bot==20.4
requests==2.31.0
